MaskGIT-code/requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.9.0.80
Pillow==10.2.0
Requests==2.31.0
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.61.2
utils==1.0.2
yacs==0.1.8
